+++ src/ReleaseNotes.fs
@@ -4,8 +4,10 @@ let AppVersion = "0.0.1"
 
 let IsPrerelease = false
 
-let AndroidVersionCode = 72
+let AndroidVersionCode = 74
 
 let Notes = """
 #### 0.0.1 - 04.01.2019
+
+* ActionBarPage for simulating the android action bar
 """