+++ RELEASE_NOTES.md
@@ -1,1 +1,3 @@
 #### 0.0.1 - 04.01.2019
+
+* ActionBarPage for simulating the android action bar